rename: 폴더 네이밍 규칙에 맞게 수정
