+++ src/layout/index.tsx
@@ -1,4 +1,4 @@
-import TabNav from '@layout/tabNav';
+import TabNav from '@layout/tab-nav';
 import Main from '@layout/main';
 
 const Layout: React.FC<React.PropsWithChildren> = ({ children }) => {
